Support for OSX frameworks.
