--- FlagKit-OSX/FlagKit-Bridging-Header.h
@@ -1,9 +1,0 @@
-//
-//  FlagKit-Bridging-Header.h
-//  FlagKit
-//
-//  Created by Yannick Heinrich on 21/10/15.
-//  Copyright © 2015 Bowtie. All rights reserved.
-//
-
-#import "NSImage+FakeInit.h"
