--- FlagKit-OSX/FlagKit-OSX.h
@@ -1,19 +1,0 @@
-//
-//  FlagKit-OSX.h
-//  FlagKit-OSX
-//
-//  Created by Yannick Heinrich on 21/10/15.
-//  Copyright © 2015 Bowtie. All rights reserved.
-//
-
-#import <Cocoa/Cocoa.h>
-
-//! Project version number for FlagKit-OSX.
-FOUNDATION_EXPORT double FlagKit_OSXVersionNumber;
-
-//! Project version string for FlagKit-OSX.
-FOUNDATION_EXPORT const unsigned char FlagKit_OSXVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <FlagKit_OSX/PublicHeader.h>
-
-#import "NSImage+FakeInit.h"
