--- FlagKit-OSX/NSImage+FakeInit.h
@@ -1,14 +1,0 @@
-//
-//  NSImage+FakeInit.h
-//  FlagKit
-//
-//  Created by Yannick Heinrich on 21/10/15.
-//  Copyright © 2015 Bowtie. All rights reserved.
-//
-
-#import <Cocoa/Cocoa.h>
-
-@interface NSImage (FakeInit)
-
-- (instancetype) initWithImage:(NSImage*) image;
-@end
