--- FlagKit-OSX/NSImage+FakeInit.m
@@ -1,17 +1,0 @@
-//
-//  NSImage+FakeInit.m
-//  FlagKit
-//
-//  Created by Yannick Heinrich on 21/10/15.
-//  Copyright © 2015 Bowtie. All rights reserved.
-//
-
-#import "NSImage+FakeInit.h"
-
-@implementation NSImage (FakeInit)
-
-- (instancetype) initWithImage:(NSImage*) image
-{
-    return image;
-}
-@end
