+++ FlagKit-iOS/FlagKit/FlagKit.h
@@ -1,0 +1,19 @@
+//
+//  FlagKit.h
+//  FlagKit
+//
+//  Created by Simon Blommegård on 21/10/15.
+//  Copyright © 2015 Bowtie. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for FlagKit.
+FOUNDATION_EXPORT double FlagKitVersionNumber;
+
+//! Project version string for FlagKit.
+FOUNDATION_EXPORT const unsigned char FlagKitVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <FlagKit/PublicHeader.h>
+
+
