+++ FlagKit.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:FlagKit-OSX/FlagKit-OSX.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:FlagKit-iOS/FlagKit-iOS.xcodeproj">
+   </FileRef>
+</Workspace>
